test: constrain Unix-only test files with //go:build unix
`GOOS=windows go vet` failed across six packages. Test files reaching for
syscall.Mkfifo, syscall.Stat_t, golang.org/x/sys/unix, and the linux||darwin
fd-based sandbox seams (secureBrowseWalkDir, sandboxFS.openReadBaseFd) had no
build constraint, so nothing told the build they were Unix-only and nothing in
CI cross-compiled them to find out.

Whole-file cases get the constraint in place: agent/execenv/grep_cancel_test.go
(renamed, linux||darwin like the seam it patches),
cmd/evener-dev/modulelint_signal_test.go (SIGUSR1/SIGUSR2), and
internal/plugins/coverage_atomic_git_fuzz_test.go (renamed; its lock coverage
scripts unix.EWOULDBLOCK/EAGAIN retries that only locks_unix.go implements).

Everywhere else only part of a file was Unix-only, so the Unix-dependent tests
and their helpers move verbatim into a sibling _unix_test.go. That keeps the
portable rest of each package cross-compiling and keeps shared helpers
(fixtureModule, buildEvenerDev, copyRepositoryFile) reachable from the untagged
files that call them — tagging those whole files broke five other files.

No test body changed: `go test -list '.*'` over the root and agent modules
produces an identical set on darwin before and after.

CI's static-build job now runs `GOOS=windows make vet` after the host vet, so an
untagged Unix-only file cannot land again.
