+++ agent/execenv/grep_cancel_unix_test.go
@@ -1,3 +1,5 @@
+//go:build linux || darwin
+
 package execenv
 
 import (
