+++ cmd/evener-dev/modulelint_signal_test.go
@@ -1,3 +1,5 @@
+//go:build unix
+
 package dev
 
 import (
