+++ internal/plugins/coverage_atomic_git_fuzz_unix_test.go
@@ -1,4 +1,4 @@
-//go:build evenerfuzz
+//go:build evenerfuzz && (linux || darwin)
 
 package plugins
